feat: add model struct

+++ domain/model/category_model.go
@@ -1,6 +1,6 @@
 package model
 
 type Category struct {
-	Id       int    `json:"id,omitempty"`
+	Id       uint64 `json:"id,omitempty"`
 	Category string `json:"category,omitempty"`
 }
